rewrite migrations

+++ database/migrations/2023_03_10_085755_create_farms_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('location');
-            $table->integer('farmer_id')->unsigned();
+            $table->bigInteger('farmer_id')->unsigned();
             $table->foreign('farmer_id')->references('id')->on('users');
             $table->timestamps();
         });
